move MockResponse to separate file
